+++ ReactCommon/fabric/components/text/paragraph/ParagraphShadowNode.h
@@ -7,11 +7,11 @@
 
 #pragma once
 
+#include <fabric/components/text/ParagraphProps.h>
+#include <fabric/components/text/TextShadowNode.h>
 #include <fabric/core/ConcreteShadowNode.h>
 #include <fabric/core/ShadowNode.h>
 #include <fabric/core/LayoutContext.h>
-#include <fabric/text/ParagraphProps.h>
-#include <fabric/text/TextShadowNode.h>
 #include <fabric/textlayoutmanager/TextLayoutManager.h>
 #include <fabric/view/ConcreteViewShadowNode.h>
 #include <folly/Optional.h>
