+++ ReactCommon/fabric/components/text/text/TextProps.h
@@ -8,10 +8,10 @@
 #pragma once
 
 #include <fabric/attributedstring/TextAttributes.h>
+#include <fabric/components/text/BaseTextProps.h>
 #include <fabric/core/Props.h>
 #include <fabric/graphics/Color.h>
 #include <fabric/graphics/Geometry.h>
-#include <fabric/text/BaseTextProps.h>
 
 namespace facebook {
 namespace react {
